feat: summarise services (#104850)
