+++ src/app/core/services/sparque-search/sparque-search.service.ts
@@ -5,14 +5,35 @@ import { Observable, map } from 'rxjs';
 import { SearchParameter, SearchResponse } from 'ish-core/models/search/search.model';
 import { SparqueSearchResponse } from 'ish-core/models/sparque-search/sparque-search.interface';
 import { SparqueSearchMapper } from 'ish-core/models/sparque-search/sparque-search.mapper';
+import { SparqueSuggestions } from 'ish-core/models/sparque-suggestion/sparque-suggestion.interface';
+import { SparqueSuggestionMapper } from 'ish-core/models/sparque-suggestion/sparque-suggestion.mapper';
+import { Suggestion } from 'ish-core/models/suggestion/suggestion.model';
 import { SearchService } from 'ish-core/services/search/search.service';
 import { SparqueApiService } from 'ish-core/services/sparque-api/sparque-api.service';
 import { URLFormParams } from 'ish-core/utils/url-form-params';
 
 @Injectable({ providedIn: 'root' })
 export class SparqueSearchService extends SearchService {
-  constructor(private sparqueApiService: SparqueApiService, private sparqueSearchMapper: SparqueSearchMapper) {
+  constructor(
+    private sparqueApiService: SparqueApiService,
+    private sparqueSuggestionMapper: SparqueSuggestionMapper,
+    private sparqueSearchMapper: SparqueSearchMapper
+  ) {
     super();
+  }
+
+  /**
+   * Searches for suggestions based on the provided search term.
+   *
+   * @param searchTerm - The term to search for suggestions.
+   * @returns An Observable emitting the search suggestions.
+   */
+  search(searchTerm: string): Observable<Suggestion> {
+    // count: maximum number of suggestions which is used individually for each type of suggestion
+    const params = new HttpParams().set('Keyword', searchTerm).set('count', '8');
+    return this.sparqueApiService
+      .get<SparqueSuggestions>(`suggestions`, { params, skipApiErrorHandling: true })
+      .pipe(map(suggestions => this.sparqueSuggestionMapper.fromData(suggestions)));
   }
 
   searchProducts(searchParams: SearchParameter): Observable<SearchResponse> {
