+++ src/app/core/store/shopping/shopping-store.spec.ts
@@ -16,12 +16,10 @@ import { CategoriesService } from 'ish-core/services/categories/categories.servi
 import { ConfigurationService } from 'ish-core/services/configuration/configuration.service';
 import { CountryService } from 'ish-core/services/country/country.service';
 import { FilterService } from 'ish-core/services/filter/filter.service';
-import { ICMSuggestionService } from 'ish-core/services/icm-suggestion/icm-suggestion.service';
 import { PricesService } from 'ish-core/services/prices/prices.service';
 import { ProductsService } from 'ish-core/services/products/products.service';
 import { PromotionsService } from 'ish-core/services/promotions/promotions.service';
 import { SearchServiceProvider } from 'ish-core/services/search/provider/search.service.provider';
-import { SuggestionServiceProvider } from 'ish-core/services/suggestion/provider/suggestion.service.provider';
 import { WarrantyService } from 'ish-core/services/warranty/warranty.service';
 import { CoreStoreModule } from 'ish-core/store/core/core-store.module';
 import { personalizationStatusDetermined } from 'ish-core/store/customer/user';
@@ -46,8 +44,6 @@ describe('Shopping Store', () => {
   let productsServiceMock: ProductsService;
   let searchServiceProviderMock: SearchServiceProvider;
   let promotionsServiceMock: PromotionsService;
-  let suggestionServiceMock: ICMSuggestionService;
-  let suggestionServiceProviderMock: SuggestionServiceProvider;
   let filterServiceMock: FilterService;
   let priceServiceMock: PricesService;
   let warrantyServiceMock: WarrantyService;
@@ -136,14 +132,10 @@ describe('Shopping Store', () => {
     when(productsServiceMock.searchProducts(anything())).thenReturn(
       of({ products: [{ sku: 'P1' }, { sku: 'P2' }] as Product[], sortableAttributes: [], total: 2 })
     );
+    when(productsServiceMock.search('some')).thenReturn(of<Suggestion>({ keywordSuggestions: ['something'] }));
 
     promotionsServiceMock = mock(PromotionsService);
     when(promotionsServiceMock.getPromotion(anything())).thenReturn(of(promotion));
-
-    suggestionServiceMock = mock(ICMSuggestionService);
-    suggestionServiceProviderMock = mock(SuggestionServiceProvider);
-    when(suggestionServiceProviderMock.get()).thenReturn(instance(suggestionServiceMock));
-    when(suggestionServiceMock.search('some')).thenReturn(of<Suggestion>({ keywordSuggestions: ['something'] }));
 
     filterServiceMock = mock(FilterService);
     when(filterServiceMock.getFilterForSearch(anything())).thenReturn(of({} as FilterNavigation));
@@ -199,7 +191,6 @@ describe('Shopping Store', () => {
         { provide: ProductsService, useFactory: () => instance(productsServiceMock) },
         { provide: PromotionsService, useFactory: () => instance(promotionsServiceMock) },
         { provide: SearchServiceProvider, useFactory: () => instance(searchServiceProviderMock) },
-        { provide: SuggestionServiceProvider, useFactory: () => instance(suggestionServiceProviderMock) },
         { provide: WarrantyService, useFactory: () => instance(warrantyServiceMock) },
         provideStoreSnapshots(),
         SelectedProductContextFacade,
